added export statement on each module, except main.js
